+++ __mocks__/packagejson.ts
@@ -3,7 +3,8 @@ module.exports = {
         name: 'atlascode',
         publisher: 'atlassian',
         rovoDev: {
-            version: '1.0.0'
+            version: '1.0.0',
+            version_bby: '1.0.0-bby'
         }
     }
 }